+++ lib/rspeed/differ.rb
@@ -13,45 +13,40 @@ module RSpeed
             data << "#{file}:#{index + 1}" if /^it/.match?(item.gsub(/\s+/, ''))
           end
         end
-
-        RSpeed::Logger.log(self, __method__, data)
       end
     end
 
     def actual_data(files:, result:)
-      result
-        .select { |item| files.include?(item[:file]) }
-        .tap { |data| RSpeed::Logger.log(self, __method__, data) }
+      result.select { |item| files.include?(item[:file]) }
     end
 
     def added_data(files:, result:)
-      (files - result.map { |item| item[:file] })
-        .map { |file| { file: file, time: 0 } }
-        .tap { |data| RSpeed::Logger.log(self, __method__, data) }
+      (files - result.map { |item| item[:file] }).map { |file| { file: file, time: 0.0 } }
     end
 
     def diff
-      files  = actual_files
-      result = RSpeed::Database.result.uniq { |item| item[:file] }
-      actual = actual_data(files: files, result: result) + added_data(files: files, result: result)
+      files   = actual_files
+      result  = RSpeed::Database.result.uniq { |item| item[:file] }
+      added   = added_data(files: files, result: result)
+      actual  = actual_data(files: files, result: result) + added_data(files: files, result: result)
+      removed = removed_data(files: files, result: result)
 
       {
-        actual:  actual.tap { |data| RSpeed::Logger.log(self, __method__, data) },
-        added:  '--',
-        removed: removed_data(files: files, result: result),
+        actual_files:  actual,
+        actual_time:   sum_time(data: actual),
+        added_files:   added,
+        added_time:    sum_time(data: added),
+        removed_files: removed,
+        removed_time:  sum_time(data: removed),
       }
     end
 
     def removed_data(files:, result:)
-      result
-        .reject { |item| files.include?(item[:file]) }
-        .tap { |data| RSpeed::Logger.log(self, __method__, data) }
+      result.reject { |item| files.include?(item[:file]) }
     end
 
-    def removed_time(data:)
-      data
-        .sum { |item| item[:time].to_f }
-        .tap { |result| RSpeed::Logger.log(self, __method__, result) }
+    def sum_time(data:)
+      data.sum { |item| item[:time].to_f }
     end
   end
 end